Document FastCdrReader optimization ideas
